feat: 모달 버튼 분리 및 카테고리 선택자 추가
