+++ src/components/modal/Modal.ts
@@ -1,4 +1,6 @@
 import { template } from "./template";
+import SelectCategory from "./select_category/SelectCategory";
+import ModalButton from "./modal_button/ModalButton";
 
 function Modal() {
   const render = () => {
@@ -11,35 +13,11 @@ function Modal() {
       mainContainer.appendChild(modal);
     }
 
-    console.log(modal);
-    console.log(mainContainer);
-    submitHandler(modal);
-    cancelHandler(modal);
+    // Elements
+    const form = document.getElementsByClassName("modal-form")[0];
+    SelectCategory().render(form);
+    ModalButton().render(modal, form);
     dimmerClickHandler(modal);
-  };
-
-  const submitHandler = (modal: HTMLElement) => {
-    const submitButton = document.getElementsByClassName("button--primary")[0];
-    console.log(submitButton);
-
-    submitButton.addEventListener("click", (event) => {
-      event.preventDefault();
-      modal.classList.remove("modal--open");
-      console.log("submit");
-      // 추가로 수행할 작업이 있다면 이 곳에 추가
-    });
-  };
-
-  const cancelHandler = (modal: HTMLElement) => {
-    const cancelButton =
-      document.getElementsByClassName("button--secondary")[0];
-    console.log(cancelButton);
-
-    cancelButton.addEventListener("click", (event) => {
-      event.preventDefault();
-      modal.classList.remove("modal--open");
-      console.log("remove");
-    });
   };
 
   const dimmerClickHandler = (modal: HTMLElement) => {
