+++ src/components/modal/modal_button/template.ts
@@ -1,0 +1,2 @@
+export const modalButtonTemplate =
+  '<div class="button-container"><button type="button"class="button button--secondary text-caption">취소하기</button type=><button class="button button--primary text-caption">추가하기</button></div>';
