+++ src/components/modal/select_category/SelectCategory.ts
@@ -1,5 +1,10 @@
+import { template } from "./template";
+
 function SelectCategory() {
-  const render = () => {};
+  const render = (form: Element) => {
+    form.innerHTML += template;
+    console.log("m", form);
+  };
 
   return {
     render,
