+++ src/components/modal/template.ts
@@ -1,2 +1,2 @@
 export const template =
-  '<div class="modal-backdrop"></div><div class="modal-container"><h2 class="modal-title text-title">새로운 음식점</h2><form><div class="button-container"><button type="button" class="button button--secondary text-caption">취소하기</button><button class="button button--primary text-caption">추가하기</button></div></form></div>';
+  '<div class="modal-backdrop"></div><div class="modal-container"><h2 class="modal-title text-title">새로운 음식점</h2><form class="modal-form"></form></div>';
